requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.3.0
click==8.1.8
discord.py==2.5.2
dotenv==0.9.9
emoji==2.14.1
frozenlist==1.5.0
idna==3.10
joblib==1.4.2
multidict==6.1.0
mysql-connector-python==9.3.0
mysqlx-connector-python==9.2.0
nltk==3.9.1
propcache==0.3.0
protobuf==4.25.3
python-dotenv==1.0.1
regex==2024.11.6
ruff==0.10.0
supervisor==4.2.5
tqdm==4.67.1
yarl==1.18.3
